apps/trello-mcp/requirements.txt
mcp[server]==1.13.1
python-dotenv==1.0.0
requests==2.31.0
python-trello==0.0.4